refactor: change names few files
